change rev field in Signature to u32 and use strotoul to extract the value. Cleanup some dead code/comments

+++ src/detect.h
@@ -433,7 +433,7 @@ typedef struct Signature_ {
 
     uint32_t id;  /**< sid, set by the 'sid' rule keyword */
     uint32_t gid; /**< generator id */
-    uint8_t rev;
+    uint32_t rev;
 
     /** classification id **/
     uint8_t class;
